Cleanup some warnings
<==== Mercurial data ====>
Branch: release_2.3.1_review

+++ common/src/plugins/resource/test_camera/testcamera_resource_searcher.cpp
@@ -48,7 +48,7 @@ bool QnTestCameraResourceSearcher::updateSocketList()
 void QnTestCameraResourceSearcher::sendBroadcast()
 {
     for (const DiscoveryInfo& info: m_sockList)
-        info.sock->sendTo(TestCamConst::TEST_CAMERA_FIND_MSG, strlen(TestCamConst::TEST_CAMERA_FIND_MSG), BROADCAST_ADDRESS, TestCamConst::DISCOVERY_PORT);
+        info.sock->sendTo(TestCamConst::TEST_CAMERA_FIND_MSG, static_cast<unsigned int>(strlen(TestCamConst::TEST_CAMERA_FIND_MSG)), BROADCAST_ADDRESS, TestCamConst::DISCOVERY_PORT);
 }
 
 QnResourceList QnTestCameraResourceSearcher::findResources(void)
